Refactor dos nomes mundipagg para gateway
