+++ lib/gateway/AntiFraud/anti_fraud_analysis_result.rb
