+++ lib/gateway/AntiFraud/query_sale_anti_fraud_analysis_history_data.rb
