+++ lib/gateway/BaseResponse.rb
