+++ lib/gateway/BoletoTransaction/boleto_transaction.rb
