+++ lib/gateway/BoletoTransaction/boleto_transaction_data.rb
