+++ lib/gateway/BoletoTransaction/boleto_transaction_options.rb
