+++ lib/gateway/BoletoTransaction/boleto_transaction_report_file.rb
