+++ lib/gateway/BoletoTransaction/boleto_transaction_result.rb
