+++ lib/gateway/CreditCardTransaction/credit_card_transaction.rb
