+++ lib/gateway/CreditCardTransaction/credit_card_transaction_data.rb
