+++ lib/gateway/CreditCardTransaction/manage_credit_card_transaction.rb
