+++ lib/gateway/CreditCardTransaction/retry_sale_credit_card_transaction.rb
