+++ lib/gateway/ErrorItem.rb
