+++ lib/gateway/ErrorReport.rb
