+++ lib/gateway/Merchant/merchant.rb
