+++ lib/gateway/Parsers/boleto_transaction_parser.rb
