+++ lib/gateway/Parsers/credit_card_transaction_parser.rb
