+++ lib/gateway/Parsers/header_parser.rb
