+++ lib/gateway/Parsers/trailer_parser.rb
