+++ lib/gateway/Person/buyer.rb
