+++ lib/gateway/Recurrency/recurrency.rb
