+++ lib/gateway/Sale/create_sale_request.rb
