+++ lib/gateway/Sale/create_sale_response.rb
