+++ lib/gateway/Sale/manage_sale_request.rb
