+++ lib/gateway/Sale/manage_sale_response.rb
