+++ lib/gateway/Sale/query_sale_request.rb
