+++ lib/gateway/Sale/query_sale_response.rb
