+++ lib/gateway/Sale/retry_sale_options.rb
