+++ lib/gateway/Sale/retry_sale_response.rb
