+++ lib/gateway/Sale/sale_data.rb
