+++ lib/gateway/SalesOption.rb
