+++ lib/gateway/ShoppingCart/shopping_cart.rb
