+++ lib/gateway/ShoppingCart/shopping_cart_item.rb
