+++ lib/gateway/Trailer.rb
