+++ lib/gateway/address.rb
