+++ lib/gateway/header.rb
