+++ lib/gateway/post_notification.rb
