+++ lib/gateway/transaction_report_file.rb
