+++ lib/mundipagg_api.rb
@@ -7,64 +7,64 @@ require 'nori'
 require 'rspec'
 require 'rexml/document'
 
-require_relative 'mundipagg/address'
-require_relative 'mundipagg/Address/billing_address'
-require_relative 'mundipagg/Address/buyer_address'
-require_relative 'mundipagg/Address/delivery_address'
+require_relative 'gateway/address'
+require_relative 'gateway/Address/billing_address'
+require_relative 'gateway/Address/buyer_address'
+require_relative 'gateway/Address/delivery_address'
 
-require_relative 'mundipagg/AntiFraud/anti_fraud_analysis_result'
-require_relative 'mundipagg/AntiFraud/query_sale_anti_fraud_analysis_data'
-require_relative 'mundipagg/AntiFraud/query_sale_anti_fraud_analysis_history_data'
+require_relative 'gateway/AntiFraud/anti_fraud_analysis_result'
+require_relative 'gateway/AntiFraud/query_sale_anti_fraud_analysis_data'
+require_relative 'gateway/AntiFraud/query_sale_anti_fraud_analysis_history_data'
 
-require_relative 'mundipagg/BoletoTransaction/boleto_transaction'
-require_relative 'mundipagg/BoletoTransaction/boleto_transaction_data'
-require_relative 'mundipagg/BoletoTransaction/boleto_transaction_options'
-require_relative 'mundipagg/BoletoTransaction/boleto_transaction_result'
-require_relative 'mundipagg/BoletoTransaction/boleto_transaction_report_file'
+require_relative 'gateway/BoletoTransaction/boleto_transaction'
+require_relative 'gateway/BoletoTransaction/boleto_transaction_data'
+require_relative 'gateway/BoletoTransaction/boleto_transaction_options'
+require_relative 'gateway/BoletoTransaction/boleto_transaction_result'
+require_relative 'gateway/BoletoTransaction/boleto_transaction_report_file'
 
-require_relative 'mundipagg/CreditCardTransaction/credit_card'
-require_relative 'mundipagg/CreditCardTransaction/credit_card_transaction'
-require_relative 'mundipagg/CreditCardTransaction/credit_card_transaction_data'
-require_relative 'mundipagg/CreditCardTransaction/credit_card_transaction_options'
-require_relative 'mundipagg/CreditCardTransaction/manage_credit_card_transaction'
-require_relative 'mundipagg/CreditCardTransaction/retry_sale_credit_card_transaction'
-require_relative 'mundipagg/CreditCardTransaction/credit_card_transaction_report_file'
+require_relative 'gateway/CreditCardTransaction/credit_card'
+require_relative 'gateway/CreditCardTransaction/credit_card_transaction'
+require_relative 'gateway/CreditCardTransaction/credit_card_transaction_data'
+require_relative 'gateway/CreditCardTransaction/credit_card_transaction_options'
+require_relative 'gateway/CreditCardTransaction/manage_credit_card_transaction'
+require_relative 'gateway/CreditCardTransaction/retry_sale_credit_card_transaction'
+require_relative 'gateway/CreditCardTransaction/credit_card_transaction_report_file'
 
-require_relative 'mundipagg/InstantBuy/credit_card_data'
-require_relative 'mundipagg/InstantBuy/get_instant_buy_data_response'
+require_relative 'gateway/InstantBuy/credit_card_data'
+require_relative 'gateway/InstantBuy/get_instant_buy_data_response'
 
-require_relative 'mundipagg/Merchant/merchant'
+require_relative 'gateway/Merchant/merchant'
 
-require_relative 'mundipagg/Order/order'
-require_relative 'mundipagg/Order/order_transaction_report_file'
+require_relative 'gateway/Order/order'
+require_relative 'gateway/Order/order_transaction_report_file'
 
-require_relative 'mundipagg/OnlineDebit/online_debit_transaction_report_file'
+require_relative 'gateway/OnlineDebit/online_debit_transaction_report_file'
 
-require_relative 'mundipagg/Parsers/boleto_transaction_parser'
-require_relative 'mundipagg/Parsers/credit_card_transaction_parser'
-require_relative 'mundipagg/Parsers/online_debit_transaction_parser'
-require_relative 'mundipagg/Parsers/header_parser'
-require_relative 'mundipagg/Parsers/trailer_parser'
+require_relative 'gateway/Parsers/boleto_transaction_parser'
+require_relative 'gateway/Parsers/credit_card_transaction_parser'
+require_relative 'gateway/Parsers/online_debit_transaction_parser'
+require_relative 'gateway/Parsers/header_parser'
+require_relative 'gateway/Parsers/trailer_parser'
 
-require_relative 'mundipagg/Person/buyer'
-require_relative 'mundipagg/Person/person'
+require_relative 'gateway/Person/buyer'
+require_relative 'gateway/Person/person'
 
-require_relative 'mundipagg/Recurrency/recurrency'
+require_relative 'gateway/Recurrency/recurrency'
 
-require_relative 'mundipagg/Sale/create_sale_request'
-require_relative 'mundipagg/Sale/manage_sale_request'
-require_relative 'mundipagg/Sale/query_sale_request'
-require_relative 'mundipagg/Sale/request_data'
-require_relative 'mundipagg/Sale/retry_sale_options'
-require_relative 'mundipagg/Sale/retry_sale_request'
-require_relative 'mundipagg/Sale/sale_data'
+require_relative 'gateway/Sale/create_sale_request'
+require_relative 'gateway/Sale/manage_sale_request'
+require_relative 'gateway/Sale/query_sale_request'
+require_relative 'gateway/Sale/request_data'
+require_relative 'gateway/Sale/retry_sale_options'
+require_relative 'gateway/Sale/retry_sale_request'
+require_relative 'gateway/Sale/sale_data'
 
-require_relative 'mundipagg/ShoppingCart/shopping_cart'
-require_relative 'mundipagg/ShoppingCart/shopping_cart_item'
+require_relative 'gateway/ShoppingCart/shopping_cart'
+require_relative 'gateway/ShoppingCart/shopping_cart_item'
 
-require_relative 'mundipagg/SalesOption'
-require_relative 'mundipagg/post_notification'
-require_relative 'mundipagg/MundipaggApi'
-require_relative 'mundipagg/header'
-require_relative 'mundipagg/Trailer'
-require_relative 'mundipagg/transaction_report_file'
+require_relative 'gateway/SalesOption'
+require_relative 'gateway/post_notification'
+require_relative 'gateway/MundipaggApi'
+require_relative 'gateway/header'
+require_relative 'gateway/Trailer'
+require_relative 'gateway/transaction_report_file'
